refactor: extract analytics/usage to layered architecture
* Create api/analytics/{link,org,usage}.rs from router.rs and api/analytics.rs
* Create repositories/analytics_repository.rs
* Create services/analytics_service.rs (gating logic + time range parsing + tests)
