+++ src/api/analytics/mod.rs
@@ -1,0 +1,6 @@
+/// Analytics API handlers
+///
+/// Org-level analytics, per-link analytics, and usage endpoints.
+pub mod link;
+pub mod org;
+pub mod usage;
