+++ src/router.rs
@@ -3,7 +3,7 @@ use crate::db;
 use crate::kv;
 use crate::middleware::{RateLimitConfig, RateLimiter};
 use crate::models::{
-    LinkAnalyticsResponse, PaginatedResponse, PaginationMeta, Tier, TimeRange,
+    PaginatedResponse, PaginationMeta, Tier,
     link::{CreateLinkRequest, Link, LinkStatus, UpdateLinkRequest},
 };
 use crate::repositories::tag_repository::{TagRepository, validate_and_normalize_tags};
@@ -894,160 +894,6 @@ pub async fn handle_get_link_by_code(req: Request, ctx: RouteContext<()>) -> Res
         Some(link) => Response::from_json(&link),
         None => Response::error("Link not found", 404),
     }
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/links/{id}/analytics",
-    tag = "Links",
-    summary = "Get link analytics",
-    description = "Returns click analytics for a single link. The time range is capped according to the organization's tier retention window (7 days for Free, 365 days for Pro, unlimited for Business/Unlimited)",
-    params(
-        ("id" = String, Path, description = "Link ID"),
-        ("days" = Option<i64>, Query, description = "Number of days to look back (default: 7)"),
-        ("start" = Option<i64>, Query, description = "Unix timestamp range start (alternative to days)"),
-        ("end" = Option<i64>, Query, description = "Unix timestamp range end"),
-    ),
-    responses(
-        (status = 200, description = "Analytics data for the link"),
-        (status = 401, description = "Unauthorized"),
-        (status = 404, description = "Link not found"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_get_link_analytics(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    // Authenticate request
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let org_id = &user_ctx.org_id;
-
-    let link_id = ctx
-        .param("id")
-        .ok_or_else(|| Error::RustError("Missing link ID".to_string()))?;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Verify link exists and belongs to org
-    let link = match db::get_link_by_id(&db, link_id, org_id).await? {
-        Some(link) => link,
-        None => return Response::error("Link not found", 404),
-    };
-
-    // Parse time range from query parameters
-    // Support both new format (TimeRange enum) and legacy format (start/end timestamps)
-    let url = req.url()?;
-    let query = url.query().unwrap_or("");
-
-    // Try to parse as new TimeRange format first
-    let time_range = if let Ok(time_range_str) = extract_query_param(query, "time_range") {
-        // New format: JSON TimeRange object
-        serde_json::from_str::<TimeRange>(&time_range_str)
-            .map_err(|e| Error::RustError(format!("Invalid time_range parameter: {}", e)))?
-    } else if let Ok(days_str) = extract_query_param(query, "days") {
-        // Simple days parameter (e.g., ?days=7)
-        let days = days_str.parse::<i64>().unwrap_or(7);
-        TimeRange::Days { value: days }
-    } else {
-        // Legacy format: start/end timestamps for backward compatibility
-        let now = crate::models::analytics::now_timestamp();
-
-        let start_legacy = query
-            .split('&')
-            .find(|s| s.starts_with("start="))
-            .and_then(|s| s.split('=').nth(1))
-            .and_then(|s| s.parse().ok())
-            .unwrap_or_else(|| now - 7 * 24 * 60 * 60); // Default: 7 days ago
-
-        let end_legacy = query
-            .split('&')
-            .find(|s| s.starts_with("end="))
-            .and_then(|s| s.split('=').nth(1))
-            .and_then(|s| s.parse().ok())
-            .unwrap_or(now);
-
-        TimeRange::Custom {
-            start: start_legacy,
-            end: end_legacy,
-        }
-    };
-
-    // Calculate timestamps using backend logic (eliminates clock skew)
-    let (mut start, end) = time_range.calculate_timestamps();
-
-    // Check tier-based analytics limits from billing account
-    let org = db::get_org_by_id(&db, org_id)
-        .await?
-        .ok_or_else(|| Error::RustError("Organization not found".to_string()))?;
-
-    // Get tier from billing account (all orgs should have billing accounts after migration)
-    let tier = if let Some(ref billing_account_id) = org.billing_account_id {
-        db::get_billing_account(&db, billing_account_id)
-            .await?
-            .and_then(|ba| Tier::from_str_value(&ba.tier))
-            .unwrap_or(Tier::Free)
-    } else {
-        Tier::Free
-    };
-
-    let mut analytics_gated = false;
-    let mut gated_reason = None;
-
-    let limits = tier.limits();
-
-    // Enforce analytics retention window — clamp start date
-    if let Some(retention_days) = limits.analytics_retention_days {
-        let now = crate::models::analytics::now_timestamp();
-        let retention_start = now - retention_days * 24 * 60 * 60;
-        if start < retention_start {
-            start = retention_start;
-            if !analytics_gated {
-                analytics_gated = true;
-                gated_reason = Some("retention_limited".to_string());
-            }
-        }
-    }
-
-    // If analytics are gated, return empty data with gating info
-    if analytics_gated {
-        let response = LinkAnalyticsResponse {
-            link,
-            total_clicks_in_range: 0,
-            clicks_over_time: vec![],
-            top_referrers: vec![],
-            top_countries: vec![],
-            top_user_agents: vec![],
-            analytics_gated: Some(true),
-            gated_reason,
-        };
-        return Response::from_json(&response);
-    }
-
-    // Run analytics queries sequentially (D1 limitation)
-    let total_clicks_in_range =
-        db::get_link_total_clicks_in_range(&db, link_id, org_id, start, end).await?;
-    let clicks_over_time = db::get_link_clicks_over_time(&db, link_id, org_id, start, end).await?;
-    let top_referrers = db::get_link_top_referrers(&db, link_id, org_id, start, end, 10).await?;
-    let top_countries = db::get_link_top_countries(&db, link_id, org_id, start, end, 10).await?;
-    let top_user_agents =
-        db::get_link_top_user_agents(&db, link_id, org_id, start, end, 20).await?;
-
-    let response = LinkAnalyticsResponse {
-        link,
-        total_clicks_in_range,
-        clicks_over_time,
-        top_referrers,
-        top_countries,
-        top_user_agents,
-        analytics_gated: None,
-        gated_reason: None,
-    };
-
-    Response::from_json(&response)
 }
 
 #[utoipa::path(
@@ -2137,87 +1983,6 @@ pub async fn handle_oauth_callback(req: Request, ctx: RouteContext<()>) -> Resul
     headers.append("Set-Cookie", &refresh_cookie)?;
 
     Ok(Response::empty()?.with_status(302).with_headers(headers))
-}
-
-#[utoipa::path(
-    get,
-    path = "/api/usage",
-    tag = "Usage",
-    summary = "Get current usage",
-    description = "Returns the authenticated organization's tier, feature limits, current monthly link usage, tag count, and the date/time of the next monthly counter reset",
-    responses(
-        (status = 200, description = "Usage and limits for the current org"),
-        (status = 401, description = "Unauthorized"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_get_usage(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let org_id = &user_ctx.org_id;
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-    let org = db::get_org_by_id(&db, org_id)
-        .await?
-        .ok_or_else(|| Error::RustError("Organization not found".to_string()))?;
-
-    // Get billing account for usage tracking
-    let billing_account_id = org
-        .billing_account_id
-        .as_ref()
-        .ok_or_else(|| Error::RustError("Organization has no billing account".to_string()))?;
-    let billing_account = db::get_billing_account(&db, billing_account_id)
-        .await?
-        .ok_or_else(|| Error::RustError("Billing account not found".to_string()))?;
-
-    let tier = Tier::from_str_value(&billing_account.tier).unwrap_or(Tier::Free);
-    let limits = tier.limits();
-
-    // Use billing account monthly counter for efficiency
-    let now = chrono::Utc::now();
-    let year_month = format!("{}-{:02}", now.year(), now.month());
-    let links_created_this_month =
-        db::get_monthly_counter_for_billing_account(&db, &billing_account.id, &year_month).await?;
-
-    // Get tag count for the billing account
-    let tags_count = TagRepository::new()
-        .count_distinct_tags_for_billing_account(&db, &billing_account.id)
-        .await?;
-
-    // Calculate next reset time (first day of next month at midnight UTC)
-    let now = chrono::Utc::now();
-    let next_reset = chrono::Utc
-        .with_ymd_and_hms(now.year(), now.month() + 1, 1, 0, 0, 0)
-        .single()
-        .unwrap_or_else(chrono::Utc::now);
-    let next_reset_timestamp = next_reset.timestamp();
-
-    let usage = serde_json::json!({
-        "tier": tier.as_str(),
-        "limits": {
-            "max_links_per_month": limits.max_links_per_month,
-            "analytics_retention_days": limits.analytics_retention_days,
-            "allow_custom_short_code": limits.allow_custom_short_code,
-            "allow_utm_parameters": limits.allow_utm_parameters,
-            "allow_query_forwarding": limits.allow_query_forwarding,
-            "max_tags": limits.max_tags,
-        },
-        "usage": {
-            "links_created_this_month": links_created_this_month,
-            "tags_count": tags_count,
-        },
-        "next_reset": {
-            "utc": next_reset.to_rfc3339(),
-            "timestamp": next_reset_timestamp,
-        }
-    });
-
-    Response::from_json(&usage)
 }
 
 #[utoipa::path(
