+++ src/services/analytics_service.rs
@@ -1,8 +1,8 @@
-use crate::auth;
-use crate::db;
+/// Analytics Service
+///
+/// Business logic for analytics gating and time range parsing.
+/// Moved from api/analytics.rs to the services layer.
 use crate::models::{Tier, TimeRange};
-use worker::d1::D1Database;
-use worker::*;
 
 /// Analytics gating result
 #[derive(Debug, Clone)]
@@ -85,92 +85,6 @@ pub fn parse_time_range_from_query_with_now(query: &str, now: i64) -> TimeRange 
     }
 }
 
-#[utoipa::path(
-    get,
-    path = "/api/analytics/org",
-    tag = "Analytics",
-    summary = "Get org-level analytics",
-    description = "Returns aggregate click analytics for the entire organization. Includes total clicks, unique links clicked, clicks over time, top links, referrers, countries, and user agents. The time range is capped by tier retention (7 days Free, 365 days Pro, unlimited Business/Unlimited)",
-    params(
-        ("days" = Option<i64>, Query, description = "Number of days to look back (default: 7)"),
-        ("start" = Option<i64>, Query, description = "Unix timestamp range start (alternative to days)"),
-        ("end" = Option<i64>, Query, description = "Unix timestamp range end"),
-    ),
-    responses(
-        (status = 200, description = "Org analytics response with clicks, top links, referrers, countries, and user agents"),
-        (status = 401, description = "Unauthorized"),
-        (status = 404, description = "Organization not found"),
-    ),
-    security(
-        ("Bearer" = []),
-        ("session_cookie" = [])
-    )
-)]
-pub async fn handle_get_org_analytics(req: Request, ctx: RouteContext<()>) -> Result<Response> {
-    let user_ctx = match auth::authenticate_request(&req, &ctx).await {
-        Ok(ctx) => ctx,
-        Err(e) => return Ok(e.into_response()),
-    };
-    let org_id = &user_ctx.org_id;
-
-    let url = req.url()?;
-    let query = url.query().unwrap_or("");
-
-    // Parse time range: ?days=N, ?start=UNIX&end=UNIX
-    let time_range = parse_time_range_from_query(query);
-
-    let (mut start, end) = time_range.calculate_timestamps();
-
-    let db = ctx.env.get_binding::<D1Database>("rushomon")?;
-
-    // Resolve tier from billing account
-    let org = db::get_org_by_id(&db, org_id)
-        .await?
-        .ok_or_else(|| Error::RustError("Organization not found".to_string()))?;
-
-    let tier = if let Some(ref billing_account_id) = org.billing_account_id {
-        db::get_billing_account(&db, billing_account_id)
-            .await?
-            .and_then(|ba| Tier::from_str_value(&ba.tier))
-            .unwrap_or(Tier::Free)
-    } else {
-        Tier::Free
-    };
-
-    let now = crate::models::analytics::now_timestamp();
-    let gating_result = apply_analytics_gating(tier, start, end, now);
-
-    // Use the adjusted start date for queries
-    start = gating_result.adjusted_start;
-
-    // Run queries sequentially (D1 limitation)
-    let total_clicks = db::get_org_total_clicks_in_range(&db, org_id, start, end).await?;
-    let unique_links_clicked = db::get_org_unique_links_clicked(&db, org_id, start, end).await?;
-    let clicks_over_time = db::get_org_clicks_over_time(&db, org_id, start, end).await?;
-    let top_links = db::get_org_top_links(&db, org_id, start, end, 10).await?;
-    let top_referrers = db::get_org_top_referrers(&db, org_id, start, end, 10).await?;
-    let top_countries = db::get_org_top_countries(&db, org_id, start, end, 10).await?;
-    let top_user_agents = db::get_org_top_user_agents(&db, org_id, start, end, 20).await?;
-
-    let response = crate::models::analytics::OrgAnalyticsResponse {
-        total_clicks,
-        unique_links_clicked,
-        clicks_over_time,
-        top_links,
-        top_referrers,
-        top_countries,
-        top_user_agents,
-        analytics_gated: if gating_result.gated {
-            Some(true)
-        } else {
-            None
-        },
-        gated_reason: gating_result.reason,
-    };
-
-    Response::from_json(&response)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -180,21 +94,18 @@ mod tests {
 
     #[test]
     fn test_parse_time_range_days_parameter() {
-        // Test valid days parameter
         let time_range = parse_time_range_from_query_with_now("days=30", TEST_NOW);
         match time_range {
             TimeRange::Days { value } => assert_eq!(value, 30),
             _ => panic!("Expected Days variant"),
         }
 
-        // Test days parameter with other params
         let time_range = parse_time_range_from_query_with_now("days=90&other=value", TEST_NOW);
         match time_range {
             TimeRange::Days { value } => assert_eq!(value, 90),
             _ => panic!("Expected Days variant"),
         }
 
-        // Test days=0 (all time)
         let time_range = parse_time_range_from_query_with_now("days=0", TEST_NOW);
         match time_range {
             TimeRange::Days { value } => assert_eq!(value, 0),
@@ -204,9 +115,8 @@ mod tests {
 
     #[test]
     fn test_parse_time_range_custom_parameters() {
-        // Test valid custom range
-        let start = 1640995200; // 2022-01-01
-        let end = 1643673600; // 2022-02-01
+        let start = 1640995200;
+        let end = 1643673600;
         let time_range =
             parse_time_range_from_query_with_now(&format!("start={}&end={}", start, end), TEST_NOW);
 
@@ -218,7 +128,6 @@ mod tests {
             _ => panic!("Expected Custom variant"),
         }
 
-        // Test custom range with other params
         let time_range = parse_time_range_from_query_with_now(
             &format!("other=value&start={}&end={}&more=data", start, end),
             TEST_NOW,
@@ -235,13 +144,12 @@ mod tests {
 
     #[test]
     fn test_parse_time_range_missing_params_defaults() {
-        // Test empty query string
         let time_range = parse_time_range_from_query_with_now("", TEST_NOW);
 
         match time_range {
             TimeRange::Custom { start, end } => {
                 assert_eq!(end, TEST_NOW);
-                assert_eq!(start, TEST_NOW - 7 * 24 * 60 * 60); // 7 days ago
+                assert_eq!(start, TEST_NOW - 7 * 24 * 60 * 60);
             }
             _ => panic!("Expected Custom variant"),
         }
@@ -249,31 +157,27 @@ mod tests {
 
     #[test]
     fn test_parse_time_range_invalid_days_fallback() {
-        // Test non-numeric days parameter
         let time_range = parse_time_range_from_query_with_now("days=invalid", TEST_NOW);
         match time_range {
-            TimeRange::Days { value } => assert_eq!(value, 7), // Should fallback to 7
+            TimeRange::Days { value } => assert_eq!(value, 7),
             _ => panic!("Expected Days variant"),
         }
 
-        // Test partially numeric days parameter
         let time_range = parse_time_range_from_query_with_now("days=30abc", TEST_NOW);
         match time_range {
-            TimeRange::Days { value } => assert_eq!(value, 7), // Should fallback to 7
+            TimeRange::Days { value } => assert_eq!(value, 7),
             _ => panic!("Expected Days variant"),
         }
 
-        // Test empty days parameter
         let time_range = parse_time_range_from_query_with_now("days=", TEST_NOW);
         match time_range {
-            TimeRange::Days { value } => assert_eq!(value, 7), // Should fallback to 7
+            TimeRange::Days { value } => assert_eq!(value, 7),
             _ => panic!("Expected Days variant"),
         }
     }
 
     #[test]
     fn test_parse_time_range_mixed_parameters_priority() {
-        // Test that days parameter takes priority over custom range
         let time_range =
             parse_time_range_from_query_with_now("days=30&start=123&end=456", TEST_NOW);
         match time_range {
@@ -281,7 +185,6 @@ mod tests {
             _ => panic!("Expected Days variant (days should take priority)"),
         }
 
-        // Test that days parameter takes priority even in different order
         let time_range =
             parse_time_range_from_query_with_now("start=123&days=90&end=456", TEST_NOW);
         match time_range {
@@ -292,24 +195,22 @@ mod tests {
 
     #[test]
     fn test_parse_time_range_partial_custom_parameters() {
-        // Test only start parameter (missing end)
         let time_range = parse_time_range_from_query_with_now("start=1640995200", TEST_NOW);
 
         match time_range {
             TimeRange::Custom { start, end } => {
                 assert_eq!(start, 1640995200);
-                assert_eq!(end, TEST_NOW); // Should default to now
+                assert_eq!(end, TEST_NOW);
             }
             _ => panic!("Expected Custom variant"),
         }
 
-        // Test only end parameter (missing start)
         let time_range = parse_time_range_from_query_with_now("end=1643673600", TEST_NOW);
 
         match time_range {
             TimeRange::Custom { start, end } => {
                 assert_eq!(end, 1643673600);
-                assert_eq!(start, TEST_NOW - 7 * 24 * 60 * 60); // Should default to 7 days ago
+                assert_eq!(start, TEST_NOW - 7 * 24 * 60 * 60);
             }
             _ => panic!("Expected Custom variant"),
         }
@@ -317,38 +218,35 @@ mod tests {
 
     #[test]
     fn test_parse_time_range_invalid_custom_parameters() {
-        // Test invalid start parameter
         let time_range =
             parse_time_range_from_query_with_now("start=invalid&end=1643673600", TEST_NOW);
 
         match time_range {
             TimeRange::Custom { start, end } => {
-                assert_eq!(start, TEST_NOW - 7 * 24 * 60 * 60); // Should default to 7 days ago
+                assert_eq!(start, TEST_NOW - 7 * 24 * 60 * 60);
                 assert_eq!(end, 1643673600);
             }
             _ => panic!("Expected Custom variant"),
         }
 
-        // Test invalid end parameter
         let time_range =
             parse_time_range_from_query_with_now("start=1640995200&end=invalid", TEST_NOW);
 
         match time_range {
             TimeRange::Custom { start, end } => {
                 assert_eq!(start, 1640995200);
-                assert_eq!(end, TEST_NOW); // Should default to now
+                assert_eq!(end, TEST_NOW);
             }
             _ => panic!("Expected Custom variant"),
         }
 
-        // Test both invalid parameters
         let time_range =
             parse_time_range_from_query_with_now("start=invalid&end=invalid", TEST_NOW);
 
         match time_range {
             TimeRange::Custom { start, end } => {
-                assert_eq!(start, TEST_NOW - 7 * 24 * 60 * 60); // Should default to 7 days ago
-                assert_eq!(end, TEST_NOW); // Should default to now
+                assert_eq!(start, TEST_NOW - 7 * 24 * 60 * 60);
+                assert_eq!(end, TEST_NOW);
             }
             _ => panic!("Expected Custom variant"),
         }
@@ -356,21 +254,18 @@ mod tests {
 
     #[test]
     fn test_parse_time_range_edge_cases() {
-        // Test negative days
         let time_range = parse_time_range_from_query_with_now("days=-7", TEST_NOW);
         match time_range {
             TimeRange::Days { value } => assert_eq!(value, -7),
             _ => panic!("Expected Days variant"),
         }
 
-        // Test very large days
         let time_range = parse_time_range_from_query_with_now("days=3650", TEST_NOW);
         match time_range {
             TimeRange::Days { value } => assert_eq!(value, 3650),
             _ => panic!("Expected Days variant"),
         }
 
-        // Test zero timestamps
         let time_range = parse_time_range_from_query_with_now("start=0&end=0", TEST_NOW);
         match time_range {
             TimeRange::Custom { start, end } => {
@@ -380,14 +275,12 @@ mod tests {
             _ => panic!("Expected Custom variant"),
         }
 
-        // Test malformed query string
         let time_range = parse_time_range_from_query_with_now("&days=30&&&", TEST_NOW);
         match time_range {
             TimeRange::Days { value } => assert_eq!(value, 30),
             _ => panic!("Expected Days variant"),
         }
 
-        // Test query with equals signs in values (should still work)
         let time_range = parse_time_range_from_query_with_now("days=30&other=value=test", TEST_NOW);
         match time_range {
             TimeRange::Days { value } => assert_eq!(value, 30),
@@ -397,14 +290,13 @@ mod tests {
 
     #[test]
     fn test_parse_time_range_parameter_order_independence() {
-        // Test different parameter orders
         let queries = [
             "days=30",
             "days=30&other=value",
             "other=value&days=30",
-            "days=30&start=123&end=456", // days should take priority
-            "start=123&days=30&end=456", // days should take priority
-            "start=123&end=456&days=30", // days should take priority
+            "days=30&start=123&end=456",
+            "start=123&days=30&end=456",
+            "start=123&end=456&days=30",
         ];
 
         for query in queries.iter() {
@@ -418,24 +310,21 @@ mod tests {
 
     #[test]
     fn test_parse_time_range_url_encoded_characters() {
-        // Test URL-encoded characters (though our simple parser doesn't decode them)
         let time_range = parse_time_range_from_query_with_now("days=30%20", TEST_NOW);
         match time_range {
-            TimeRange::Days { value } => assert_eq!(value, 7), // Should fallback due to %20
+            TimeRange::Days { value } => assert_eq!(value, 7),
             _ => panic!("Expected Days variant"),
         }
     }
 
     #[test]
     fn test_parse_time_range_multiple_same_parameters() {
-        // Test multiple days parameters (first one should be used)
         let time_range = parse_time_range_from_query_with_now("days=30&days=90", TEST_NOW);
         match time_range {
             TimeRange::Days { value } => assert_eq!(value, 30),
             _ => panic!("Expected Days variant"),
         }
 
-        // Test multiple start parameters (first one should be used)
         let time_range =
             parse_time_range_from_query_with_now("start=123&start=456&end=789", TEST_NOW);
 
@@ -452,74 +341,68 @@ mod tests {
     #[test]
     fn test_analytics_gating_free_tier() {
         let now = TEST_NOW;
-        let start_8_days_ago = now - 8 * 24 * 60 * 60; // 8 days ago
-        let start_6_days_ago = now - 6 * 24 * 60 * 60; // 6 days ago
+        let start_8_days_ago = now - 8 * 24 * 60 * 60;
+        let start_6_days_ago = now - 6 * 24 * 60 * 60;
 
-        // Test free tier (7-day retention)
         let result = apply_analytics_gating(Tier::Free, start_8_days_ago, now, now);
 
         assert!(result.gated);
         assert_eq!(result.reason, Some("retention_limited".to_string()));
-        assert_eq!(result.adjusted_start, now - 7 * 24 * 60 * 60); // Clamped to 7 days ago
+        assert_eq!(result.adjusted_start, now - 7 * 24 * 60 * 60);
         assert_eq!(result.original_start, start_8_days_ago);
 
-        // Test free tier with valid range (no gating)
         let result = apply_analytics_gating(Tier::Free, start_6_days_ago, now, now);
 
         assert!(!result.gated);
         assert_eq!(result.reason, None);
-        assert_eq!(result.adjusted_start, start_6_days_ago); // Unchanged
+        assert_eq!(result.adjusted_start, start_6_days_ago);
         assert_eq!(result.original_start, start_6_days_ago);
     }
 
     #[test]
     fn test_analytics_gating_pro_tier() {
         let now = TEST_NOW;
-        let start_400_days_ago = now - 400 * 24 * 60 * 60; // 400 days ago
-        let start_300_days_ago = now - 300 * 24 * 60 * 60; // 300 days ago
+        let start_400_days_ago = now - 400 * 24 * 60 * 60;
+        let start_300_days_ago = now - 300 * 24 * 60 * 60;
 
-        // Test pro tier (365-day retention)
         let result = apply_analytics_gating(Tier::Pro, start_400_days_ago, now, now);
 
         assert!(result.gated);
         assert_eq!(result.reason, Some("retention_limited".to_string()));
-        assert_eq!(result.adjusted_start, now - 365 * 24 * 60 * 60); // Clamped to 365 days ago
+        assert_eq!(result.adjusted_start, now - 365 * 24 * 60 * 60);
         assert_eq!(result.original_start, start_400_days_ago);
 
-        // Test pro tier with valid range (no gating)
         let result = apply_analytics_gating(Tier::Pro, start_300_days_ago, now, now);
 
         assert!(!result.gated);
         assert_eq!(result.reason, None);
-        assert_eq!(result.adjusted_start, start_300_days_ago); // Unchanged
+        assert_eq!(result.adjusted_start, start_300_days_ago);
         assert_eq!(result.original_start, start_300_days_ago);
     }
 
     #[test]
     fn test_analytics_gating_business_tier() {
         let now = TEST_NOW;
-        let start_1000_days_ago = now - 1000 * 24 * 60 * 60; // 1000 days ago
+        let start_1000_days_ago = now - 1000 * 24 * 60 * 60;
 
-        // Test business tier (unlimited retention)
         let result = apply_analytics_gating(Tier::Business, start_1000_days_ago, now, now);
 
         assert!(!result.gated);
         assert_eq!(result.reason, None);
-        assert_eq!(result.adjusted_start, start_1000_days_ago); // Unchanged
+        assert_eq!(result.adjusted_start, start_1000_days_ago);
         assert_eq!(result.original_start, start_1000_days_ago);
     }
 
     #[test]
     fn test_analytics_gating_unlimited_tier() {
         let now = TEST_NOW;
-        let start_5000_days_ago = now - 5000 * 24 * 60 * 60; // 5000 days ago
+        let start_5000_days_ago = now - 5000 * 24 * 60 * 60;
 
-        // Test unlimited tier (unlimited retention)
         let result = apply_analytics_gating(Tier::Unlimited, start_5000_days_ago, now, now);
 
         assert!(!result.gated);
         assert_eq!(result.reason, None);
-        assert_eq!(result.adjusted_start, start_5000_days_ago); // Unchanged
+        assert_eq!(result.adjusted_start, start_5000_days_ago);
         assert_eq!(result.original_start, start_5000_days_ago);
     }
 
@@ -527,31 +410,27 @@ mod tests {
     fn test_analytics_gating_edge_cases() {
         let now = TEST_NOW;
 
-        // Test exactly at retention limit (free tier, 7 days)
         let exactly_7_days_ago = now - 7 * 24 * 60 * 60;
         let result = apply_analytics_gating(Tier::Free, exactly_7_days_ago, now, now);
 
-        assert!(!result.gated); // Should not be gated at exactly the limit
+        assert!(!result.gated);
         assert_eq!(result.reason, None);
         assert_eq!(result.adjusted_start, exactly_7_days_ago);
 
-        // Test exactly at retention limit (pro tier, 365 days)
         let exactly_365_days_ago = now - 365 * 24 * 60 * 60;
         let result = apply_analytics_gating(Tier::Pro, exactly_365_days_ago, now, now);
 
-        assert!(!result.gated); // Should not be gated at exactly the limit
+        assert!(!result.gated);
         assert_eq!(result.reason, None);
         assert_eq!(result.adjusted_start, exactly_365_days_ago);
 
-        // Test start date in the future (should not be gated)
-        let future_start = now + 24 * 60 * 60; // 1 day in future
+        let future_start = now + 24 * 60 * 60;
         let result = apply_analytics_gating(Tier::Free, future_start, now, now);
 
         assert!(!result.gated);
         assert_eq!(result.reason, None);
         assert_eq!(result.adjusted_start, future_start);
 
-        // Test start date equals now (should not be gated)
         let result = apply_analytics_gating(Tier::Free, now, now, now);
 
         assert!(!result.gated);
@@ -563,7 +442,6 @@ mod tests {
     fn test_analytics_gating_zero_and_negative_ranges() {
         let now = TEST_NOW;
 
-        // Test start = 0 (epoch) with free tier
         let result = apply_analytics_gating(Tier::Free, 0, now, now);
 
         assert!(result.gated);
@@ -571,8 +449,7 @@ mod tests {
         assert_eq!(result.adjusted_start, now - 7 * 24 * 60 * 60);
         assert_eq!(result.original_start, 0);
 
-        // Test negative start (before epoch) with free tier
-        let result = apply_analytics_gating(Tier::Free, -86400, now, now); // 1 day before epoch
+        let result = apply_analytics_gating(Tier::Free, -86400, now, now);
 
         assert!(result.gated);
         assert_eq!(result.reason, Some("retention_limited".to_string()));
@@ -583,17 +460,15 @@ mod tests {
     #[test]
     fn test_analytics_gating_result_structure() {
         let now = TEST_NOW;
-        let start = now - 10 * 24 * 60 * 60; // 10 days ago
+        let start = now - 10 * 24 * 60 * 60;
 
         let result = apply_analytics_gating(Tier::Free, start, now, now);
 
-        // Test that all fields are properly set
         assert_eq!(result.gated, true);
         assert_eq!(result.reason, Some("retention_limited".to_string()));
-        assert!(result.adjusted_start > result.original_start); // Should be clamped forward
+        assert!(result.adjusted_start > result.original_start);
         assert_eq!(result.original_start, start);
 
-        // Test cloning works
         let cloned = result.clone();
         assert_eq!(cloned.gated, result.gated);
         assert_eq!(cloned.reason, result.reason);
@@ -604,9 +479,8 @@ mod tests {
     #[test]
     fn test_analytics_gating_consistency_across_tiers() {
         let now = TEST_NOW;
-        let start = now - 10 * 24 * 60 * 60; // 10 days ago
+        let start = now - 10 * 24 * 60 * 60;
 
-        // Test that unlimited tiers never gate
         let business_result = apply_analytics_gating(Tier::Business, start, now, now);
         let unlimited_result = apply_analytics_gating(Tier::Unlimited, start, now, now);
 
@@ -617,11 +491,10 @@ mod tests {
         assert_eq!(business_result.adjusted_start, start);
         assert_eq!(unlimited_result.adjusted_start, start);
 
-        // Test that limited tiers gate appropriately
         let free_result = apply_analytics_gating(Tier::Free, start, now, now);
         let pro_result = apply_analytics_gating(Tier::Pro, start, now, now);
 
-        assert!(free_result.gated); // 10 days > 7-day limit
-        assert!(!pro_result.gated); // 10 days < 365-day limit
+        assert!(free_result.gated);
+        assert!(!pro_result.gated);
     }
 }
